+++ sdk/spring/spring-cloud-azure-appconfiguration-config/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 5.16.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 5.15.0 (2024-08-07)
 
 Please refer to [spring/CHANGELOG.md](https://github.com/Azure/azure-sdk-for-java/tree/main/sdk/spring/CHANGELOG.md#5150-2024-08-07) for more details.